+++ server/controllers/shared/caseListController.ts
@@ -2,7 +2,7 @@ import type { Request, Response, TypedRequestHandler } from 'express'
 
 import { assessPaths } from '../../paths'
 import type { ReferralService } from '../../services'
-import { PathUtils, ReferralUtils, TypeUtils } from '../../utils'
+import { CaseListUtils, PathUtils, TypeUtils } from '../../utils'
 import type { QueryParam } from '@accredited-programmes/ui'
 
 export default class CaseListController {
@@ -35,16 +35,16 @@ export default class CaseListController {
       const { activeCaseLoadId, username } = res.locals.user
 
       const paginatedReferralSummaries = await this.referralService.getReferralSummaries(username, activeCaseLoadId, {
-        audience: ReferralUtils.uiToApiAudienceQueryParam(audience),
-        status: ReferralUtils.uiToApiStatusQueryParam(status),
+        audience: CaseListUtils.uiToApiAudienceQueryParam(audience),
+        status: CaseListUtils.uiToApiStatusQueryParam(status),
       })
 
       return res.render('referrals/caseList/show', {
         action: assessPaths.caseList.filter({}),
-        audienceSelectItems: ReferralUtils.audienceSelectItems(audience),
+        audienceSelectItems: CaseListUtils.audienceSelectItems(audience),
         pageHeading: 'My referrals',
-        referralStatusSelectItems: ReferralUtils.statusSelectItems(status),
-        tableRows: ReferralUtils.caseListTableRows(paginatedReferralSummaries.content),
+        referralStatusSelectItems: CaseListUtils.statusSelectItems(status),
+        tableRows: CaseListUtils.caseListTableRows(paginatedReferralSummaries.content),
       })
     }
   }
