+++ server/paths/index.ts
@@ -1,9 +1,18 @@
 import apiPaths from './api'
-import assessPaths from './assess'
+import assessPaths, { assessPathBase } from './assess'
 import findPaths from './find'
 import prisonApiPaths from './prisonApi'
 import prisonRegisterApiPaths from './prisonRegisterApi'
 import prisonerSearchPaths from './prisonerSearch'
 import referPaths from './refer'
 
-export { apiPaths, assessPaths, findPaths, prisonApiPaths, prisonRegisterApiPaths, prisonerSearchPaths, referPaths }
+export {
+  apiPaths,
+  assessPathBase,
+  assessPaths,
+  findPaths,
+  prisonApiPaths,
+  prisonRegisterApiPaths,
+  prisonerSearchPaths,
+  referPaths,
+}
