+++ server/utils/index.ts
@@ -11,7 +11,9 @@ import OffenceUtils from './offenceUtils'
 import OrganisationUtils from './organisationUtils'
 import PathUtils from './pathUtils'
 import PersonUtils from './personUtils'
-import ReferralUtils from './referralUtils'
+import CaseListUtils from './referrals/caseListUtils'
+import NewReferralUtils from './referrals/newReferralUtils'
+import ShowReferralUtils from './referrals/showReferralUtils'
 import RouteUtils from './routeUtils'
 import SentenceInformationUtils from './sentenceInformationUtils'
 import StringUtils from './stringUtils'
@@ -19,17 +21,19 @@ import TypeUtils from './typeUtils'
 import UserUtils from './userUtils'
 
 export {
+  CaseListUtils,
   CourseParticipationUtils,
   CourseUtils,
   DateUtils,
   FormUtils,
+  NewReferralUtils,
   OffenceUtils,
   OrganisationUtils,
   PathUtils,
   PersonUtils,
-  ReferralUtils,
   RouteUtils,
   SentenceInformationUtils,
+  ShowReferralUtils,
   StringUtils,
   TypeUtils,
   UserUtils,
